Defining the Results container class
